Reorganized the files into packages and sub-packages
