+++ src/main/java/edu/clemson/cs/r2jt/errors/ErrorHandler2.java
@@ -13,7 +13,7 @@
 package edu.clemson.cs.r2jt.errors;
 
 import edu.clemson.cs.r2jt.ResolveCompiler;
-import edu.clemson.cs.r2jt.init.CompileEnvironment2;
+import edu.clemson.cs.r2jt.init2.model.CompileEnvironment2;
 
 /**
  * TODO: Description for this class
