+++ src/main/java/edu/clemson/cs/r2jt/init2/ResolveCompiler2.java
@@ -10,12 +10,12 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package edu.clemson.cs.r2jt;
+package edu.clemson.cs.r2jt.init2;
 
 import edu.clemson.cs.r2jt.archiving.Archiver;
 import edu.clemson.cs.r2jt.congruenceclassprover.CongruenceClassProver;
 import edu.clemson.cs.r2jt.congruenceclassprover.SMTProver;
-import edu.clemson.cs.r2jt.init.CompileEnvironment2;
+import edu.clemson.cs.r2jt.init2.model.CompileEnvironment2;
 import edu.clemson.cs.r2jt.misc.Flag;
 import edu.clemson.cs.r2jt.misc.FlagDependencies;
 import edu.clemson.cs.r2jt.misc.FlagDependencyException;
