+++ src/main/java/edu/clemson/cs/r2jt/init2/controller/Controller2.java
@@ -10,10 +10,11 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package edu.clemson.cs.r2jt.init;
+package edu.clemson.cs.r2jt.init2.controller;
 
-import edu.clemson.cs.r2jt.compilereport.CompileReport2;
+import edu.clemson.cs.r2jt.init2.model.CompileReport2;
 import edu.clemson.cs.r2jt.errors.ErrorHandler2;
+import edu.clemson.cs.r2jt.init2.model.CompileEnvironment2;
 
 /**
  * A manager for the target file of a compilation.
